+++ scripts/roop_globals.py
@@ -3,7 +3,9 @@ import os
 
 MODELS_DIR = os.path.abspath(os.path.join("models","roop"))
 ANALYZER_DIR = os.path.abspath(os.path.join(MODELS_DIR, "analysers"))
-VERSION_FLAG = "v1.0.1"
+FACE_PARSER_DIR = os.path.abspath(os.path.join(MODELS_DIR, "parser"))
+
+VERSION_FLAG = "v1.1.0"
 EXTENSION_PATH=os.path.join("extensions","sd-webui-roop")
 SD_CONVERT_SCORE = 0.7