+++ scripts/roop_postprocessing/i2i_pp.py
@@ -8,26 +8,28 @@ import numpy as np
 from modules import shared
 from scripts.roop_utils import imgutils
 from modules import shared, processing, codeformer_model
-
+from pprint import pformat
 from modules.processing import (StableDiffusionProcessingImg2Img)
 from enum import Enum
+from scripts.roop_postprocessing.postprocessing_options import PostProcessingOptions, InpaintingWhen
+from modules import sd_models
 
 from scripts.roop_swapping import swapper
 
 
-def img2img_diffusion(img : Image.Image, inpainting_prompt : str, inpainting_denoising_strength : float = 0.1, inpainting_negative_prompt : str="", inpainting_steps : int = 20, inpainting_sampler : str ="Euler") -> Image.Image :
-    if inpainting_denoising_strength == 0  :
+def img2img_diffusion(img : Image.Image, pp: PostProcessingOptions) -> Image.Image :
+    if pp.inpainting_denoising_strengh == 0  :
         return img
 
     try :
         logger.info(
 f"""Inpainting face
-Sampler : {inpainting_sampler}
-inpainting_denoising_strength : {inpainting_denoising_strength}
-inpainting_steps : {inpainting_steps}
+Sampler : {pp.inpainting_sampler}
+inpainting_denoising_strength : {pp.inpainting_denoising_strengh}
+inpainting_steps : {pp.inpainting_steps}
 """
 )
-        if not isinstance(inpainting_sampler, str) :
+        if not isinstance(pp.inpainting_sampler, str) :
             inpainting_sampler = "Euler"
 
         logger.info("send faces to image to image")
@@ -37,12 +39,36 @@ inpainting_steps : {inpainting_steps}
             for face in faces:
                 bbox =face.bbox.astype(int)
                 mask = imgutils.create_mask(img, bbox)
-                prompt = inpainting_prompt.replace("[gender]", "man" if face["gender"] == 1 else "woman")
-                negative_prompt = inpainting_negative_prompt.replace("[gender]", "man" if face["gender"] == 1 else "woman")
+                prompt = pp.inpainting_prompt.replace("[gender]", "man" if face["gender"] == 1 else "woman")
+                negative_prompt = pp.inpainting_negative_prompt.replace("[gender]", "man" if face["gender"] == 1 else "woman")
                 logger.info("Denoising prompt : %s", prompt)
-                logger.info("Denoising strenght : %s", inpainting_denoising_strength)                
-                i2i_p = StableDiffusionProcessingImg2Img([img],sampler_name=inpainting_sampler, do_not_save_samples=True, steps =inpainting_steps, width = img.width, inpainting_fill=1, inpaint_full_res= True, height = img.height, mask=mask, prompt = prompt,negative_prompt=negative_prompt, denoising_strength=inpainting_denoising_strength)
+                logger.info("Denoising strenght : %s", pp.inpainting_denoising_strengh)                
+                
+                i2i_kwargs = {"sampler_name" :pp.inpainting_sampler,
+                        "do_not_save_samples":True, 
+                        "steps" :pp.inpainting_steps,
+                        "width" : img.width,
+                        "inpainting_fill":1,
+                        "inpaint_full_res":True,
+                        "height" : img.height,
+                        "mask": mask,
+                        "prompt" : prompt,
+                        "negative_prompt" :negative_prompt,
+                        "denoising_strength" :pp.inpainting_denoising_strengh}
+                current_model_checkpoint = shared.opts.sd_model_checkpoint
+                if pp.inpainting_model and pp.inpainting_model != "Current" :
+                    # Change checkpoint
+                    shared.opts.sd_model_checkpoint = pp.inpainting_model
+                    sd_models.select_checkpoint
+                    sd_models.load_model()
+                i2i_p = StableDiffusionProcessingImg2Img([img], **i2i_kwargs)
                 i2i_processed = processing.process_images(i2i_p)
+                if pp.inpainting_model and pp.inpainting_model != "Current" :
+                    # Restore checkpoint
+                    shared.opts.sd_model_checkpoint = current_model_checkpoint
+                    sd_models.select_checkpoint
+                    sd_models.load_model()
+
                 images = i2i_processed.images
                 if len(images) > 0 :
                     img = images[0]