+++ scripts/roop_swapping/upscaled_inswapper.py
@@ -1,29 +1,61 @@
 import time
-import numpy as np
-import onnxruntime
-import cv2
-import onnx
-from onnx import numpy_helper
-from  insightface.utils import face_align
-from  insightface.model_zoo.inswapper import INSwapper
-from modules import scripts, shared, processing
-from modules.face_restoration import FaceRestoration
-from modules.upscaler import UpscalerData
-from PIL import Image
-from scripts.roop_logging import logger
-from scripts.roop_utils.imgutils import cv2_to_pil, pil_to_cv2
-from modules import processing
-from modules.shared import cmd_opts, opts, state
-from scripts.roop_postprocessing.postprocessing_options import PostProcessingOptions
-from scripts.roop_postprocessing import upscaling
 
+import cv2
+import numpy as np
+import onnx
+import onnxruntime
+from insightface.model_zoo.inswapper import INSwapper
+from insightface.utils import face_align
+from modules import codeformer_model, processing, scripts, shared
+from modules.face_restoration import FaceRestoration
+from modules.shared import cmd_opts, opts, state
+from modules.upscaler import UpscalerData
+from onnx import numpy_helper
+from PIL import Image
+
+from scripts.roop_logging import logger
+from scripts.roop_postprocessing import upscaling
+from scripts.roop_postprocessing.postprocessing_options import \
+    PostProcessingOptions
+from scripts.roop_utils.imgutils import cv2_to_pil, pil_to_cv2
+from scripts.roop_swapping.facemask import generate_face_mask
 
 def get_upscaler() -> UpscalerData:
     for upscaler in shared.sd_upscalers:
         if upscaler.name == opts.data.get("roop_upscaled_swapper_upscaler", "LDSR"):
             return upscaler
     return None
-            
+
+def merge_images_with_mask(image1, image2, mask):
+    if image1.shape != image2.shape or image1.shape[:2] != mask.shape:
+        raise ValueError("Img should have the same shape")
+    mask = mask.astype(np.uint8)
+    masked_region = cv2.bitwise_and(image2, image2, mask=mask)
+    inverse_mask = cv2.bitwise_not(mask)
+    empty_region = cv2.bitwise_and(image1, image1, mask=inverse_mask)
+    merged_image = cv2.add(empty_region, masked_region)
+    return merged_image
+
+def erode_mask(mask, kernel_size=3, iterations=1):
+    kernel = np.ones((kernel_size, kernel_size), np.uint8)
+    eroded_mask = cv2.erode(mask, kernel, iterations=iterations)
+    return eroded_mask
+
+def apply_gaussian_blur(mask, kernel_size=(5, 5), sigma_x=0):
+    blurred_mask = cv2.GaussianBlur(mask, kernel_size, sigma_x)
+    return blurred_mask
+
+def dilate_mask(mask, kernel_size=5, iterations=1):
+    kernel = np.ones((kernel_size, kernel_size), np.uint8)
+    dilated_mask = cv2.dilate(mask, kernel, iterations=iterations)
+    return dilated_mask
+
+def get_face_mask(aimg,bgr_fake):
+    mask1 = generate_face_mask(aimg, device = shared.device)
+    mask2 = generate_face_mask(bgr_fake, device = shared.device)
+    mask = dilate_mask(cv2.bitwise_or(mask1,mask2))    
+    return mask
+
 
 class UpscaledINSwapper():
     def __init__(self, inswapper : INSwapper):
@@ -75,15 +107,21 @@ class UpscaledINSwapper():
                     return fake_diff
 
                 if upscale :
+
                     print("*"*80)
                     print(f"Upscaled inswapper using {opts.data.get('roop_upscaled_swapper_upscaler', 'LDSR')}")
                     print("*"*80)
+                                    
                     k = 4
                     aimg, M = face_align.norm_crop2(img, target_face.kps, self.input_size[0]*k)                
                    
                     # upscale and restore face :
                     bgr_fake = self.super_resolution(bgr_fake, k)
                     
+                    if opts.data.get("roop_upscaled_improved_mask", True) :
+                        mask = get_face_mask(aimg,bgr_fake)
+                        bgr_fake = merge_images_with_mask(aimg, bgr_fake,mask)
+
                     # compute fake_diff before sharpen and color correction (better result)
                     fake_diff = compute_diff(bgr_fake, aimg)
 
@@ -99,6 +137,8 @@ class UpscaledINSwapper():
                         correction = processing.setup_color_correction(cv2_to_pil(aimg))
                         bgr_fake_pil = processing.apply_color_correction(correction, cv2_to_pil(bgr_fake))
                         bgr_fake = pil_to_cv2(bgr_fake_pil)
+
+
                 else :
                     fake_diff = compute_diff(bgr_fake, aimg)